[PATCH] usb hcd u64 warning fix
drivers/usb/core/hcd.c:1689: warning: long long unsigned int format, u64 arg (arg 6)
drivers/usb/core/hcd.c:1695: warning: long long unsigned int format, u64 arg (arg 5)

We must not assume that u64 is implemented as `unsigned long long'.  On ppc64
(for example) it is `unsigned long'.

+++ drivers/usb/core/hcd.c
@@ -1683,13 +1683,15 @@ int usb_add_hcd(struct usb_hcd *hcd,
 		hcd->irq = irqnum;
 		dev_info(hcd->self.controller, "irq %s, %s 0x%08llx\n", bufp,
 				(hcd->driver->flags & HCD_MEMORY) ?
-					"io mem" : "io base", hcd->rsrc_start);
+					"io mem" : "io base",
+					(unsigned long long)hcd->rsrc_start);
 	} else {
 		hcd->irq = -1;
 		if (hcd->rsrc_start)
 			dev_info(hcd->self.controller, "%s 0x%08llx\n",
 					(hcd->driver->flags & HCD_MEMORY) ?
-					"io mem" : "io base", hcd->rsrc_start);
+					"io mem" : "io base",
+					(unsigned long long)hcd->rsrc_start);
 	}
 
 	if ((retval = hcd->driver->start(hcd)) < 0) {
